move control plane resize command under osdctl cluster resize
